requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.1.4
django-admin-volt==1.0.4
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-sendgrid-v5==1.2.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==5.1.0
joblib==1.2.0
Markdown==3.4.1
numpy==1.24.1
pandas==1.5.3
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyngrok==5.2.1
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.1
scipy==1.10.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
threadpoolctl==3.1.0
urllib3==1.26.14
yenepay==0.5.0a0
zipp==3.11.0
